# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "description": "\n    Autentica usuário com email e senha, retornando tokens de acesso JWT.\n    \n    **Funcionalidades:**\n    - Validação de credenciais\n    - Geração de access_token (válido por 1h)\n    - Geração de refresh_token (válido por 7 dias)\n    - Registro de sessão com IP e User-Agent\n    - Controle de tentativas de login\n    ",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              },
              "examples": {
                "admin": {
                  "summary": "Login de Administrador",
                  "description": "Exemplo de login com usuário administrador",
                  "value": {
                    "email": "admin@neuralcontent.com",
                    "password": "admin123"
                  }
                },
                "user": {
                  "summary": "Login de Usuário",
                  "description": "Exemplo de login com usuário comum",
                  "value": {
                    "email": "usuario@exemplo.com",
                    "password": "senha123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                },
                "example": {
                  "success": true,
                  "message": "Login realizado com sucesso",
                  "data": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "token_type": "Bearer",
                    "expires_in": 3600,
                    "user": {
                      "id": 1,
                      "firstName": "João",
                      "lastName": "Silva",
                      "email": "usuario@exemplo.com",
                      "role": "user",
                      "isActive": true,
                      "avatar": null,
                      "createdAt": "2025-01-01T00:00:00.000Z"
                    }
                  },
                  "timestamp": "2025-07-17T20:45:12.023Z"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas ou usuário inativo",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Email ou senha inválidos",
                  "error": "INVALID_CREDENTIALS",
                  "timestamp": "2025-07-17T20:45:12.023Z"
                }
              }
            }
          },
          "422": {
            "description": "Dados de entrada inválidos",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Dados de entrada inválidos",
                  "errors": [
                    {
                      "field": "email",
                      "message": "Email deve ter um formato válido",
                      "value": "email-invalido"
                    },
                    {
                      "field": "password",
                      "message": "Senha deve ter pelo menos 6 caracteres",
                      "value": "123"
                    }
                  ],
                  "timestamp": "2025-07-17T20:45:12.023Z"
                }
              }
            }
          }
        },
        "summary": "Login do usuário",
        "tags": [
          "🔐 Autenticação"
        ]
      }
    },
    "/v1/auth/register": {
      "post": {
        "description": "\n    Cria uma nova conta de usuário no sistema.\n    \n    **Funcionalidades:**\n    - Validação de unicidade do email\n    - Criptografia segura da senha (bcrypt)\n    - Geração automática de tokens JWT\n    - Criação de perfil inicial\n    - Envio de email de boas-vindas (opcional)\n    ",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              },
              "examples": {
                "basic": {
                  "summary": "Registro Básico",
                  "description": "Exemplo básico de registro de usuário",
                  "value": {
                    "firstName": "João",
                    "lastName": "Silva",
                    "email": "joao.silva@exemplo.com",
                    "password": "senha123",
                    "confirmPassword": "senha123"
                  }
                },
                "complete": {
                  "summary": "Registro Completo",
                  "description": "Exemplo com todos os campos opcionais",
                  "value": {
                    "firstName": "Maria",
                    "lastName": "Santos",
                    "email": "maria.santos@exemplo.com",
                    "password": "minhasenha456",
                    "confirmPassword": "minhasenha456",
                    "acceptTerms": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                },
                "example": {
                  "success": true,
                  "message": "Usuário registrado com sucesso",
                  "data": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "token_type": "Bearer",
                    "expires_in": 3600,
                    "user": {
                      "id": 2,
                      "firstName": "João",
                      "lastName": "Silva",
                      "email": "joao.silva@exemplo.com",
                      "role": "user",
                      "isActive": true,
                      "avatar": null,
                      "emailVerifiedAt": null,
                      "createdAt": "2025-07-17T20:45:12.023Z"
                    }
                  },
                  "timestamp": "2025-07-17T20:45:12.023Z"
                }
              }
            }
          },
          "400": {
            "description": "Email já está em uso",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Este email já está registrado no sistema",
                  "error": "EMAIL_ALREADY_EXISTS",
                  "timestamp": "2025-07-17T20:45:12.023Z"
                }
              }
            }
          },
          "422": {
            "description": "Dados de entrada inválidos",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Dados de entrada inválidos",
                  "errors": [
                    {
                      "field": "email",
                      "message": "Email deve ter um formato válido",
                      "value": "email-invalido"
                    },
                    {
                      "field": "password",
                      "message": "Senha deve ter pelo menos 6 caracteres",
                      "value": "123"
                    },
                    {
                      "field": "confirmPassword",
                      "message": "Confirmação de senha não confere",
                      "value": "diferente"
                    }
                  ],
                  "timestamp": "2025-07-17T20:45:12.023Z"
                }
              }
            }
          }
        },
        "summary": "Registro de usuário",
        "tags": [
          "🔐 Autenticação"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "description": "Gera novo access token usando refresh token válido",
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token renovado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token inválido ou expirado"
          }
        },
        "summary": "Renovar token de acesso",
        "tags": [
          "🔐 Autenticação"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "description": "Invalida refresh token e encerra sessão",
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logout realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Logout realizado com sucesso"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token de acesso inválido"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout do usuário",
        "tags": [
          "🔐 Autenticação"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "description": "Retorna informações do usuário logado",
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dados do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "email": {
                      "type": "string",
                      "example": "usuario@neuralcontent.com"
                    },
                    "name": {
                      "type": "string",
                      "example": "João Silva Santos"
                    },
                    "role": {
                      "type": "string",
                      "example": "user"
                    },
                    "status": {
                      "type": "string",
                      "example": "active"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token de acesso inválido"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dados do usuário autenticado",
        "tags": [
          "🔐 Autenticação"
        ]
      }
    },
    "/v1/auth/my-verification-tokens": {
      "get": {
        "description": "Retorna todos os tokens de verificação válidos de um usuário.",
        "operationId": "AuthController_getMyVerificationTokens",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Tokens encontrados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                      },
                      "userId": {
                        "type": "string",
                        "example": "123"
                      },
                      "token": {
                        "type": "string",
                        "example": "verification_token_hash"
                      },
                      "type": {
                        "type": "string",
                        "example": "email_verification"
                      },
                      "expiresAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "isUsed": {
                        "type": "boolean",
                        "example": false
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter tokens de verificação de um usuário",
        "tags": [
          "🔐 Autenticação"
        ]
      }
    },
    "/v1/auth/invalidate-my-verification-tokens": {
      "post": {
        "description": "Invalida todos os tokens de verificação de um usuário específico.",
        "operationId": "AuthController_invalidateMyVerificationTokens",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Tokens invalidados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "affected": {
                      "type": "number",
                      "example": 3
                    },
                    "message": {
                      "type": "string",
                      "example": "Tokens invalidados com sucesso"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invalidar tokens de verificação de um usuário",
        "tags": [
          "🔐 Autenticação"
        ]
      }
    },
...
  },
  "info": {
    "title": "🧠 NeuralContent API",
    "description": "\n## API completa para plataforma de geração de conteúdo com IA\n\n### 🔐 Autenticação\nEsta API utiliza **JWT Bearer tokens**. Para autenticar suas requisições:\n```\nAuthorization: Bearer {seu_jwt_token}\n```\n\n### 📊 Rate Limiting\n- **Usuários Free:** 100 req/min\n- **Usuários Pro:** 500 req/min  \n- **Usuários Enterprise:** 1000 req/min\n- **Administradores:** Sem limite\n\n### 🔄 Versionamento\n- **Versão Atual:** v1.0.0\n- **URL Base:** `/v1`\n- **Suporte:** Versionamento semântico (SemVer)\n\n### 📈 Funcionalidades Principais\n- **Autenticação e Autorização** (JWT + Roles)\n- **Gestão de Usuários** (CRUD + Perfis)\n- **Sistema de Planos** (Free, Pro, Enterprise)\n- **Gestão de Créditos** (Compra, Consumo, Histórico)\n- **Processamento de Pagamentos** (Stripe Integration)\n- **Monitoramento e Logs** (Health Checks + Analytics)\n\n### 🛡️ Segurança\n- Criptografia de senhas (bcrypt)\n- Validação de entrada (class-validator)\n- Proteção CORS configurável\n- Headers de segurança (Helmet)\n- Rate limiting por usuário\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "NeuralContent Support",
      "url": "https://neuralcontent.com/support",
      "email": "api-support@neuralcontent.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://neuralcontent.com/terms"
  },
...
  "servers": [
    {
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "JWT token obtido via /auth/login. Formato: Bearer {token}",
        "in": "header"
      },
      "ApiKey-auth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key para integração de sistemas externos"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email do usuário",
            "example": "usuario@neuralcontent.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Senha do usuário",
            "example": "MinhaS3nh@Segura",
            "minLength": 6,
            "maxLength": 50
          },
          "rememberMe": {
            "type": "boolean",
            "description": "Manter sessão ativa por mais tempo",
            "example": false,
            "default": false
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Token de acesso JWT",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "Token para renovação de acesso",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "tokenType": {
            "type": "string",
            "description": "Tipo do token",
            "example": "Bearer",
            "default": "Bearer"
          },
          "expiresIn": {
            "type": "number",
            "description": "Tempo de expiração do access token em segundos",
            "example": 900
          },
          "expiresAt": {
            "type": "number",
            "description": "Timestamp de quando o token expira",
            "example": 1625097600
          },
          "user": {
            "type": "object",
            "description": "Dados básicos do usuário autenticado",
            "example": {
              "id": 1,
              "email": "usuario@neuralcontent.com",
              "firstName": "João",
              "lastName": "Silva Santos",
              "role": "user",
              "isEmailVerified": true
            }
          },
          "permissions": {
            "description": "Permissões específicas do usuário",
            "example": [
              "read_profile",
              "write_content",
              "manage_credits"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sessionId": {
            "type": "string",
            "description": "ID da sessão para controle",
            "example": "uuid-v4-generated"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType",
          "expiresIn",
          "expiresAt",
          "user"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "Primeiro nome do usuário",
            "example": "João",
            "minLength": 2,
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "description": "Sobrenome do usuário",
            "example": "Silva Santos",
            "minLength": 2,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "description": "Email único do usuário",
            "example": "joao.silva@neuralcontent.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Senha segura do usuário",
            "example": "MinhaS3nh@SuperSegura123",
            "minLength": 8,
            "maxLength": 50
          },
          "role": {
            "type": "string",
            "description": "Papel do usuário no sistema",
            "example": "user",
            "enum": [
              "user",
              "admin",
              "moderator"
            ],
            "default": "user"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do usuário (opcional)",
            "example": "+5511999999999"
          },
          "acceptTerms": {
            "type": "boolean",
            "description": "Aceitar termos de uso e política de privacidade",
            "example": true,
            "default": false
          },
          "acceptMarketing": {
            "type": "boolean",
            "description": "Aceitar receber comunicações de marketing",
            "example": false,
            "default": false
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token válido para renovar acesso",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
    }
  }
}